OBGM-566 Fix validation messages on pack, pick and send page
